commit final - tudo 10

+++ funcionarios/produto/excluir_produto.php
@@ -1,15 +1,25 @@
 <?php
-    $id_produto = $_GET['id_produto'];
+    $id_produto = $_POST['id_produto'];
     $conectar  = mysql_connect('localhost','root','');
     $db        = mysql_select_db('frangelato');
-    $sql       = "delete from produto WHERE
-    id_produto = '$id_produto';";
-    $resultado = mysql_query($sql);
+
+    $produtosA = array();
+
+    $produtoSelect = "SELECT id_produto_processo FROM processo_fabricacao";
+    $produtoQuery = mysql_query($produtoSelect);
+    while($produtos = mysql_fetch_array($produtoQuery)){
+        $produto = $produtos['id_produto_processo'];
+        array_push($produtosA, $produto);
+    }
+
+    if(in_array($id_produto, $produtosA)){
+        echo 'Exclusão recusada: O produto está registrado em fabricação!';
+    }
+    else{
+        $sql       = "delete from produto WHERE
+        id_produto = '$id_produto';";
+        $resultado = mysql_query($sql);
+        echo 'Excluído com Sucesso!';
+    }
 ?>
 
-<script>
-	alert('excluido com Sucesso!');
-	<?php
-		echo "location.href='main_produto.php'";
-	?>
-</script>
